Refactor test fixtures for users and base
Refactor the test fixtures for users and base by organizing them into separate modules and directories. This improves the structure and maintainability of the test code.
