+++ .github/workflows/check_code_quality.yaml
@@ -16,16 +16,16 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: "3.x"
-      # - name: Install dependencies
-      #   run: |
-      #     python -m pip install --upgrade pip
-      #     pip install black isort Pylint
-      # - name: Run isort
-      #   run: |
-      #     isort .
-      # - name: Run black
-      #   run: |
-      #     black .
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install black isort Pylint
+      - name: Run isort
+        run: |
+          isort .
+      - name: Run black
+        run: |
+          black .
 
       # - name: Run Pylint
       #   run: |
