+++ test/base_fixtures/get_test_db_session.py
@@ -1,0 +1,10 @@
+import pytest
+from src.database import SessionLocal
+
+@pytest.fixture(scope="session")
+def get_test_db_session():
+    try:
+        db = SessionLocal()
+        yield db
+    finally:
+        db.close()
