+++ test/users/fixtures/__init__.py
@@ -1,0 +1,1 @@
+from .others import *
