+++ test/users/fixtures/others/member_test_service_factory.py
@@ -1,0 +1,11 @@
+import pytest
+from test.users.members_test_service import MembersTestService
+
+
+@pytest.fixture
+def member_test_service(generate_user_data
+                       , 
+                       get_user_by_identifier, get_all_users, get_access_token,get_all_admins,get_all_secretaries):
+    return MembersTestService(generate_user_data,  
+                             get_user_by_identifier, 
+                            get_all_users, get_access_token,get_all_admins,get_all_secretaries)
